refactor: update app structure and improve metrics data handling
